% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{card_crime}
\alias{card_crime}
\title{Example Card crimes}
\format{an object of class \code{list}.}
\usage{
data(card_crime)
}
\description{
Locations package authors have used credit cards around london.
}
\examples{
plot(card_crime$latitude, card_crime$longitude, xlab = "lon", ylab = "lat")

}
\keyword{datasets}
