% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{card_source}
\alias{card_source}
\title{Example Card sources}
\format{an object of class \code{list}.}
\usage{
data(card_source)
}
\description{
Old residence and qmul longitude latitude associated with card usage across london.
}
\examples{
plot(card_crime$latitude, card_crime$longitude, xlab = "lon", ylab = "lat")
points(card_source$latitude, card_source$longitude, pch = 15, col = "blue")

}
\keyword{datasets}
