% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_rgeoprofile_loaded}
\alias{check_rgeoprofile_loaded}
\title{Check that RgeoProfile package has loaded successfully}
\usage{
check_rgeoprofile_loaded()
}
\description{
Simple function to check that RgeoProfile package has loaded 
  successfully. Prints "RgeoProfile loaded successfully!" if so.
}
