% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cholera}
\alias{Cholera}
\title{Cholera data}
\format{an object of class \code{list}.}
\usage{
data(Cholera)
}
\description{
Locations of cases of cholera from the 1854 cholera outbreak made famous by John Snow.
}
\examples{
plot(Cholera$latitude, Cholera$longitude, xlab = "lon", ylab = "lat")

}
\keyword{datasets}
