\name{Choleradata}
\alias{Choleradata}
\docType{data}
\title{
Cholera cases from John Snows classic study
}
\description{

A set of Longditude and Latitude positions of the homes of cholera victims taken from John Snows classic study.
}
\usage{data(Choleradata)}
\format{
  A data frame with 321 observations on the following 2 variables.
  \describe{
    \item{\code{V1}}{a numeric vector of longitude}
    \item{\code{V2}}{a numeric vector of latitude}
  }
}

\source{
Snow, J., Frost, W. H., & Richardson, B. W. (1936). Snow on cholera. Commonwealth Fund.
}

\examples{
data(Choleradata)

}
\keyword{datasets}
