% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desalvo.R
\docType{data}
\name{desalvo}
\alias{desalvo}
\title{Incidents from the Boston Strangler Case (Albert DeSalvo)}
\format{
A data frame with 13 rows and 5 variables.
\describe{
  \item{name}{Victim name.}
  \item{age}{Age of the victim.}
  \item{date}{Date when the incident occurred.}
  \item{lat}{The latitude of the location where the incident occurred.}
  \item{lon}{The longitude of the location where the incident occurred.}
}
}
\usage{
desalvo
}
\description{
A sample dataset of serial crimes for analysis (Boston Strangler)
}
\keyword{datasets}
