% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dist_gc}
\alias{dist_gc}
\title{Calculate pairwise great circle distance between points}
\usage{
dist_gc(x)
}
\arguments{
\item{x}{a two-column matrix or dataframe with longitude in the first column
and latitude in the second}
}
\description{
Analogue of the \code{dist()} function, but calculating great
  circle distances. Points should be input as a two-column matrix or
  dataframe with longitude in the first column and latitude in the second.
}
