\name{drootinvgamma}
\alias{drootinvgamma}

\title{
Density estimator of inverge gamma function
}
\description{
Used in clalculating the density of the inverge gamma prior on the model parameter sigma
}
\usage{
drootinvgamma(x, shape, rate, log = F)
}

\arguments{
  \item{x}{
The data
}
  \item{shape}{
Shape parameter
}
  \item{rate}{
Rate parameter
}
  \item{log}{
\code{TRUE} or \code{FALSE}, wether the inverge gamma is on the log scale
}
}

\value{

\item{output }{The inverse gamma distribution}

}

\author{
MD, Stevenson & R, Verity
}



