% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoMCMC}
\alias{geoMCMC}
\title{MCMC under Rgeoprofile model}
\usage{
geoMCMC(data, params)
}
\arguments{
\item{data}{A list of input data.}

\item{params}{A list of input parameters.}
}
\description{
This function carries out the main MCMC under the Rgeoprofile model.
}
\examples{
geoMCMC()
}

