% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoParamsCheck}
\alias{geoParamsCheck}
\title{Check parameters}
\usage{
geoParamsCheck(params, silent = FALSE)
}
\arguments{
\item{params}{a list of parameters, as defined by geoParams().}

\item{silent}{whether to report passing check to console.}
}
\description{
Check that all parameters for use in Rgeoprofile MCMC are in the correct format.
}
\examples{
myData <- geoData()
myParams <- geoParams(myData, sigma_var=1)
geoParamsCheck(myParams)
}

