% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{geoPlotCoallocation}
\alias{geoPlotCoallocation}
\title{Calculate and plot probability of coallocation}
\usage{
geoPlotCoallocation(mcmc, cols = NULL)
}
\arguments{
\item{mcmc}{TODO}

\item{cols}{TODO}
}
\description{
TODO
}
\examples{
# TODO
}
