% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoPlotGini}
\alias{geoPlotGini}
\title{Produces a Gini or Lorenz plot showing the proportion of suspect sites identified as a function of area.
Also allows an optional vector called crimeNumbers with numbers of crimes per suspect}
\usage{
geoPlotGini(hit_scores = geoReportHitscores(params, source_data = source_data,
  surface = myMCMC$surface), crimeNumbers = NULL, suspects_col = "red",
  crimes_col = "blue")
}
\description{
Produces a Gini or Lorenz plot showing the proportion of suspect sites identified as a function of area.
Also allows an optional vector called crimeNumbers with numbers of crimes per suspect
}
\examples{
geoPlotGini()
}

