% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{geoPlotLeaflet}
\alias{geoPlotLeaflet}
\title{Plot a map and overlay data and/or geoprofile via leaflet}
\usage{
geoPlotLeaflet(params = NULL, data = NULL, surface = NULL,
  source = NULL, surfaceCols = c("#F0F921FF", "#FDC926FF", "#FA9E3BFF",
  "#ED7953FF", "#D8576BFF", "#BD3786FF", "#9C179EFF", "#7301A8FF",
  "#47039FFF", "#0D0887FF"), crimeCex = 1.5, crimeCol = "red",
  sourceCex = 1.5, sourceCol = "blue", map_type = 110,
  threshold = 0.1, opacity = 0.8, colOpacity = 1, smoothing = 1,
  gpLegend = FALSE)
}
\arguments{
\item{params}{parameters list in the format defined by geoParams().}

\item{data}{data object in the format defined by geoData().}

\item{surface}{a surface to overlay onto the map, typically a geoprofile obtained from the output of geoMCMC().}

\item{source}{potential sources object in the format defined by geoDataSource().}

\item{surfaceCols}{vector of two or more colours to plot surface. Defaults to viridis palette.}

\item{crimeCex}{relative size of symbols showing crimes.}

\item{crimeCol}{colour of crime symbols.}

\item{sourceCex}{relative size of symbols showing suspect sites.}

\item{sourceCol}{colour of suspect sites symbols.}

\item{map_type}{the specific type of map to plot. See leaflet::providers for options.}

\item{threshold}{what level of the geoprofile to display}

\item{opacity}{value between 0 and 1 givin the opacity of surface colours.}

\item{colOpacity}{opacity of crime and source colours}

\item{smoothing}{smooth profile}

\item{gpLegend}{whether or not to add legend to plot.}
}
\description{
Plots geoprofile on map, with various customisable options.
}
\examples{
\dontrun{
# London example data
d <- LondonExample_crimes
s <- LondonExample_sources
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2)
m <- geoMCMC(data = d, params = p)
# produce simple map
geoPlotLeaflet(params = p, 
               data = d, 
               source = s, 
               surface = m$geoProfile,
               map_type = 110,
               crimeCol = "black", 
               crimeCex = 2, 
               sourceCol = "red", 
               sourceCex = 2)

# John Snow cholera data
d <- Cholera
s <- WaterPumps
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2)
m <- geoMCMC(data = d, params = p, lambda=0.05)
# produce simple map
geoPlotLeaflet(params = p, 
               data = d, 
               source = s, 
               surface = m$geoProfile,
               map_type = 110,
               crimeCol = "black", 
               crimeCex = 2, 
               sourceCol = "red", 
               sourceCex = 2)

# simulated data
sim <-rDPM(50, priorMean_longitude = -0.04217491, priorMean_latitude = 
51.5235505, alpha=1, sigma=1, tau=3)
d <- geoData(sim$longitude, sim $latitude)
s <- geoDataSource(sim$source_lon, sim$source_lat)
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2)
m <- geoMCMC(data = d, params = p)
# change colour palette, map type, opacity and range of geoprofile and omit legend
geoPlotLeaflet(params = p, 
               data = d, 
               source = s, 
               surface = m$geoProfile,
               map_type = 110, 
               surfaceCols = c("blue","white"), 
               crimeCol = "black", 
               crimeCex = 2, 
               sourceCol = "red", 
               sourceCex = 2,
               opacity = 0.7, 
               gpLegend = FALSE)
}
}
