% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoPlotLorenz}
\alias{geoPlotLorenz}
\title{Produce Lorenz Plot}
\usage{
geoPlotLorenz(hit_scores, crimeNumbers = NULL, suspects_col = "red",
  crimes_col = "blue")
}
\description{
Produces a Lorenz plot showing the proportion of suspect sites identified as a function of area.
Also allows an optional vector called crimeNumbers with numbers of crimes per suspect. Calculates and returns Gini coefficient.
}
\examples{
geoPlotLorenz()
}
