% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoPlotMap}
\alias{geoPlotMap}
\title{Plot a map and overlay data and/or geoprofile}
\usage{
geoPlotMap(params, data = NULL, source = NULL, surface = NULL,
  zoom = "auto", mapSource = "google", mapType = "hybrid",
  transparency = 0.6, plotContours = TRUE, breakPercent = seq(0, 100, l =
  11), contourCols = c("red", "orange", "yellow", "white"), crimeCex = 1.5,
  crimeCol = "red", crimeBorderCol = "white", crimeBorderWidth = 0.5,
  sourceCex = 1.5, sourceCol = "blue")
}
\arguments{
\item{surface}{some text}
}
\description{
Creates geoprofile plot, choosing some parameters automatically. Now plots sources too (SLC).
}
\examples{
geoQuickPlot(surface)
}

