% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoPlotMapText}
\alias{geoPlotMapText}
\title{Plot a map and overlay data and/or geoprofile}
\usage{
geoPlotMapText(params, data = NULL, source = NULL, crimeNames = NULL,
  sourceNames = NULL, surface = NULL, zoom = NULL, mapSource = "google",
  mapType = "hybrid", opacity = 0.6, plotContours = TRUE,
  breakPercent = seq(0, 100, l = 11), contourCols = c("red", "orange",
  "yellow", "white"), crimeCex = 1.5, crimeCol = "red",
  crimeBorderCol = "white", crimeBorderWidth = 0.5, sourceCex = 1.5,
  sourceCol = "blue", gpLegend = TRUE)
}
\arguments{
\item{params}{parameters list in the format defined by geoParams().}

\item{data}{data object in the format defined by geoData().}

\item{source}{potential sources object in the format defined by geoDataSource().}

\item{crimeNames}{text labels for crimes. If NULL, numbers will be used.}

\item{sourceNames}{text labels for crimes. If NULL, numbers will be used.}

\item{surface}{a surface to overlay onto the map, typically a geoprofile obtained from the output of geoMCMC().}

\item{zoom}{zoom level of map. If NULL then choose optimal zoom from params.}

\item{mapSource}{which online source to use when downloading the map. Options include Google Maps ("google"), OpenStreetMap ("osm"), Stamen Maps ("stamen") and CloudMade maps ("cloudmade").}

\item{mapType}{the specific type of map to plot. Options available are "terrain", "terrain-background", "satellite", "roadmap" and "hybrid" (google maps), "terrain", "watercolor" and "toner" (stamen maps) or a positive integer for cloudmade maps (see ?get_cloudmademap from the package ggmap for details).}

\item{opacity}{value between 0 and 1 givin the opacity of surface colours.}

\item{plotContours}{whether or not to add contours to the surface plot.}

\item{breakPercent}{vector of values between 0 and 100 describing where in the surface contours appear.}

\item{contourCols}{list of two or more colours from which to derive the contour colours.}

\item{crimeCex}{relative size of symbols showing crimes.}

\item{crimeCol}{colour of crime symbols.}

\item{crimeBorderCol}{border colour of crime symbols.}

\item{crimeBorderWidth}{width of border of crime symbols.}

\item{sourceCex}{relative size of symbols showing suspect sites.}

\item{sourceCol}{colour of suspect sites symbols.}

\item{gpLegend}{whether or not to add legend to plot.}
}
\description{
Plots geoprofile on map, with various customisable options.
}
\examples{
# John Snow cholera data
data(Cholera)
d <- geoData(Cholera[,1],Cholera[,2])
data(WaterPumps)
s <- geoDataSource(WaterPumps[,1], WaterPumps[,2])
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2)
m <- geoMCMC(data = d, params = p, lambda=0.05)
# produce simple map
geoPlotMapText(params = p, data = d, source = s, crimeNames = NULL, sourceNames = letters[1:13], surface = m$geoProfile, breakPercent = seq(0, 50, 5), mapType = "hybrid",
crimeCol = "black", crimeCex = 2, sourceCol = "red", sourceCex = 2)

}
