% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoPlotSigma}
\alias{geoPlotSigma}
\title{Plot prior and posterior distributions of sigma}
\usage{
geoPlotSigma(params, sigma)
}
\arguments{
\item{params}{A list of parameters (defines prior on sigma)}

\item{sigma}{A vector of posterior draws of sigma}
}
\description{
Plot prior and posterior distributions of sigma.
}

