% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoPlotSurface}
\alias{geoPlotSurface}
\title{Plot geoprofile}
\usage{
geoPlotSurface(longitude_midpoints = 1:ncol(surface),
  latitude_midpoints = 1:nrow(surface), surface)
}
\arguments{
\item{surface}{}
}
\description{
Produces plot of posterior surface or geoprofile
}
\examples{
geoPlotSurface(surface)
}
\keyword{bob}

