% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoQuickPlot}
\alias{geoQuickPlot}
\title{Create quick geoprofile plot}
\usage{
geoQuickPlot(params, surface = NULL, data = NULL, zoom = "auto",
  source = "google", maptype = "hybrid", breakPercent = seq(0, 100, l =
  11), contour_cols = c("red", "orange", "yellow", "white"),
  plotContours = TRUE, data_fillCol = "black", data_borderCol = "white")
}
\description{
Creates quick geoprofile plot, choosing some parameters automatically.
}
\examples{
geoQuickPlot(surface)
}

