% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoRing}
\alias{geoRing}
\title{Calculate and plot hit scores based on a ring search}
\usage{
geoRing(params, data, source, mcmc)
}
\arguments{
\item{params}{Parameters list in the format defined by geoParams().}

\item{data}{Data object in the format defined by geoData().}

\item{source}{Potential sources object in the format defined by geoDataSource().}

\item{mcmc}{mcmc object of the form produced by geoMCMC().}
}
\description{
Second attempt at ring search without using other packages.
}
\examples{
# TODO
}
