% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoRingHitscores}
\alias{geoRingHitscores}
\title{Calculate and plot hit scores based on a ring search}
\usage{
geoRingHitscores(data, source, mcmc)
}
\description{
Second attempt at ring search without using other packages. TODO - complete help for this function!
}
