% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoShapefile}
\alias{geoShapefile}
\title{Import shapefile}
\usage{
geoShapefile(fileName = NULL)
}
\arguments{
\item{fileName}{The object to be imported. Must be one of SpatialPolygonsDataFrame, SpatialLinesDataFrame or RasterLayer if it is to be used with geoMask().}
}
\description{
This function imports spatial information in the form of SpatialPolygonsDataFrame, SpatialLinesDataFrame or RasterLayer for use with geoMask().
}
\examples{
# load London boroughs by default
geoShapefile()
}
