% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoSmooth}
\alias{geoSmooth}
\title{Apply smoothing to surface}
\usage{
geoSmooth(x, y, z, bandwidth)
}
\arguments{
\item{z}{matrix on which to apply smoothing.}

\item{bandwidth}{standard deviation of Gaussian smoothing kernel}
}
\description{
Applies Gaussian smoothing to surface.
}
\examples{
geoSmooth(MCMCoutput$geoSurface)
}
\keyword{bob}

