% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_output}
\alias{get_output}
\title{Get output}
\usage{
get_output(project, name, K = NULL, type = "summary")
}
\arguments{
\item{project}{an RgeoProfile project, as produced by the function 
\code{rgeoprofile_project()}}

\item{name}{name of output to get}

\item{K}{get output for this value of K}

\item{type}{the type of output ("summary" or "raw")}
}
\description{
Get output from a project for a given value of K.
}
