% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{gini_coefficient}
\alias{gini_coefficient}
\title{Calculates the gini coefficient for the probability surface.
This is a measure of unevenness, with a value of 0 meaning perfectly
even and a value of 1 maximally uneven.}
\usage{
gini_coefficient(y_vals = myMCMC$surface)
}
\description{
Calculates the gini coefficient for the probability surface.
This is a measure of unevenness, with a value of 0 meaning perfectly
even and a value of 1 maximally uneven.
}
\examples{
gini_coefficient()
}

