% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoprofile_project.R
\name{is.rgeoprofile_project}
\alias{is.rgeoprofile_project}
\title{Determine if object is of class rgeoprofile_project}
\usage{
is.rgeoprofile_project(x)
}
\arguments{
\item{x}{object of class \code{rgeoprofile_project}}
}
\description{
Determine if object is of class rgeoprofile_project
}
