% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{kernel_smooth}
\alias{kernel_smooth}
\title{Produce a smooth surface using 2D kernel density smoothing}
\usage{
kernel_smooth(longitude, latitude, breaks_lon, breaks_lat, lambda = NULL,
  nu = 3)
}
\arguments{
\item{longitude}{longitude of input points}

\item{latitude}{latitude of input points}

\item{breaks_lon}{positions of longitude breaks}

\item{breaks_lat}{positions of latitude breaks}

\item{lambda}{bandwidth to use in posterior smoothing. If NULL then optimal 
bandwidth is chosen automatically by maximum-likelihood}

\item{nu}{degrees of freedom of student's-t kernel}
}
\description{
Takes lon/lat coordinates, bins in two dimensions and smooths using kernel 
density smoothing. Kernel densities are computed using the fast Fourier 
transform method, which is many times faster than simple summation when using
a large number of points. Each Kernel is student's-t distributed and scaled
by the bandwidth lambda. If lambda is set to \code{NULL} then the optimal
value of lambda is chosen automatically using the leave-one-out maximum
likelihood method.
}
\examples{
# TODO
}
\references{
Barnard, Etienne. "Maximum leave-one-out likelihood for kernel
  density estimation." Proceedings of the Twenty-First Annual Symposium of
  the Pattern Recognition Association of South Africa. 2010
}
