% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{latlon_to_bearing}
\alias{latlon_to_bearing}
\title{Convert lat-lon to bearing}
\usage{
latlon_to_bearing(origin_lat, origin_lon, dest_lat, dest_lon)
}
\arguments{
\item{origin_lat}{latitude of origin point}

\item{origin_lon}{longitude of origin point}

\item{dest_lat}{latitude of destination point}

\item{dest_lon}{longitude of destination point}
}
\description{
Calculate bearing and great circle distance between an origin and one or more destination points
}
