% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LondonExample_sources}
\alias{LondonExample_sources}
\title{Example London sources}
\format{an object of class \code{list}.}
\usage{
data(LondonExample_sources)
}
\description{
Dummy sources associated with the crimes in LondonExample_crimes.
}
\examples{
plot(LondonExample_crimes$latitude, LondonExample_crimes$longitude, xlab = "lon", ylab = "lat")
points(LondonExample_sources$latitude, LondonExample_sources$longitude, pch = 15, col = "blue")

}
\keyword{datasets}
