% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lonlat_to_bearing}
\alias{lonlat_to_bearing}
\title{Calculate great circle distance and bearing between coordinates}
\usage{
lonlat_to_bearing(origin_lon, origin_lat, dest_lon, dest_lat)
}
\arguments{
\item{origin_lon}{The origin longitude}

\item{origin_lat}{The origin latitude}

\item{dest_lon}{The destination longitude}

\item{dest_lat}{The destination latitude}
}
\description{
Calculate great circle distance and bearing between spatial
  coordinates.
}
\examples{
# one degree longitude should equal approximately 111km at the equator
lonlat_to_bearing(0, 0, 1, 0)
}
