% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lonlat_to_cartesian}
\alias{lonlat_to_cartesian}
\title{Convert lat/lon to cartesian coordinates}
\usage{
lonlat_to_cartesian(centre_lon, centre_lat, data_lon, data_lat)
}
\arguments{
\item{centre_lon}{The centre longitude}

\item{centre_lat}{The centre latitude}

\item{data_lon}{The data longitude}

\item{data_lat}{The data latitude}
}
\description{
Convert lat/lon coordinates to cartesian coordinates by first
  calculating great circle distance and bearing and then mapping these
  coordinates into cartesian space. This mapping is relative to the point
  {centre_lat, centre_lon}, which should be roughly at the midpoint of the
  observed data.
}
\examples{
# TODO
}
