% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{more_colours}
\alias{more_colours}
\title{Expand series of colours by interpolation}
\usage{
more_colours(n = 5, raw_cols = brewer.pal(12, "Paired"))
}
\arguments{
\item{n}{how many colours to return}

\item{raw_cols}{vector of colours to interpolate}
}
\description{
Expand a series of colours by interpolation to produce any 
  number of colours from a given series. The pattern of interpolation is 
  designed so that (n+1)th value contains the nth value plus one more colour,
  rather than being a completely different series. For example, running
  \code{more_colours(5)} and \code{more_colours(4)}, the first 4 colours will
  be shared between the two series.
}
