% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{overlay_geoprofile}
\alias{overlay_geoprofile}
\title{Add geoprofile to dynamic map}
\usage{
overlay_geoprofile(myplot, project, K = NULL, source = NULL,
  threshold = 0.1, col = col_hotcold(), opacity = 0.8)
}
\arguments{
\item{myplot}{dynamic map produced by \code{plot_map()} function}

\item{project}{an RgeoProfile project, as produced by the function 
\code{rgeoprofile_project()}}

\item{K}{which value of K to plot}

\item{source}{which source to plot. If NULL then plot combined surface}

\item{threshold}{what proportion of geoprofile to plot}

\item{col}{set of plotting colours}

\item{opacity}{opacity of geoprofile (that is not invisible due to being below threshold)}
}
\description{
Add geoprofile to dynamic map
}
