% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{overlay_points}
\alias{overlay_points}
\title{Add points to dynamic map}
\usage{
overlay_points(myplot, lon, lat, col = "black", size = 1,
  opacity = 1)
}
\arguments{
\item{myplot}{dynamic map produced by \code{plot_map()} function}

\item{lon}{longitude of points}

\item{lat}{latitude of points}

\item{col}{colour of points}

\item{size}{size of points}

\item{opacity}{opacity of points}
}
\description{
Add points to dynamic map
}
