% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{overlay_sources}
\alias{overlay_sources}
\title{Add sources to dynamic map}
\usage{
overlay_sources(myplot, lon, lat)
}
\arguments{
\item{myplot}{dynamic map produced by \code{plot_map()} function}

\item{lon}{longitude of sources}

\item{lat}{latitude of sources}
}
\description{
Add sources to dynamic map
}
