% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{perspGP}
\alias{perspGP}
\title{Perspective plot of geoprofile or raw probabilities}
\usage{
perspGP(surface, aggregate_size = 3, perspCol = c("red", "orange", "yellow",
  "white"), phiGP = 30, thetaGP = -30, surface_type = "gp")
}
\arguments{
\item{surface}{surface to plot; either the geoprofile or posteriorSurface output by geoMCMC().}

\item{aggregate_size}{the number of cells to aggregate to smooth the surface.}

\item{perspCol}{colour palette. Defaults to red/orange/yellow/white.}

\item{phiGP}{value of phi to pass to persp().}

\item{thetaGP}{value of theta to pass to persp().}

\item{surface_type}{type of surface; should be either "gp" for geoprofile or "prob" for posteriorSurface.}
}
\description{
Plots persp plot of geoprofile or posterior surface (coloured according to height), reducing matrix dimensions if necessary to avoid grid lines being too close together. NB Only works with square matrix
}
\examples{
# john snow cholera data
data(Cholera)
d <- geoData(Cholera[,1],Cholera[,2])
data(WaterPumps)
s <- geoDataSource(WaterPumps[,1], WaterPumps[,2])
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2, samples = 20000, 
chains = 10, burnin = 1000, priorMean_longitude = mean(d$longitude), 
priorMean_latitude = mean(d$latitude), guardRail = 0.1)
m <- geoMCMC(data = d, params = p)
# raw probabilities
perspGP(m$posteriorSurface, surface_type = "prob")
# geoprofile
perspGP(surface = m$geoProfile, aggregate_size = 3, surface_type = "gp")

# simulated data
sim <-rDPM(50, priorMean_longitude = -0.04217491, priorMean_latitude = 
51.5235505, alpha=1, sigma=1, tau=3)
d <- geoData(sim$longitude, sim $latitude)
s <- geoDataSource(sim$source_lon, sim$source_lat)
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2, samples = 20000, 
chains = 10, burnin = 1000, priorMean_longitude = mean(d$longitude), 
priorMean_latitude = mean(d$latitude), guardRail = 0.1)
m <- geoMCMC(data = d, params = p)
# raw probabilities
perspGP(m$posteriorSurface, surface_type = "prob")
# geoprofile
perspGP(surface = m$geoProfile, aggregate_size = 3, surface_type = "gp")
}
