% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{perspGP2}
\alias{perspGP2}
\title{Interactive perspective plot of geoprofile or raw probabilities}
\usage{
perspGP2(surface, surface_type = "gp", perspCol = c("red", "orange",
  "yellow", "white"), scale = 1)
}
\arguments{
\item{surface}{surface to plot; either the geoprofile or posteriorSurface output by geoMCMC().}

\item{surface_type}{type of surface; should be either "gp" for geoprofile or "prob" for posteriorSurface.}

\item{perspCol}{colour palette. Defaults to red/orange/yellow/white.}

\item{scale}{vertical scale of surface.}
}
\description{
Produces interactive perspective plot of geoprofile or posterior surface (coloured according to height).
}
\examples{
# John Snow cholera data
data(Cholera)
d <- geoData(Cholera[,1],Cholera[,2])
data(WaterPumps)
s <- geoDataSource(WaterPumps[,1], WaterPumps[,2])
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2)
m <- geoMCMC(data = d, params = p, lambda=0.05)
# raw probabilities
perspGP2(m$posteriorSurface, surface_type = "prob")
# geoprofile
perspGP2(m$geoProfile, surface_type = "gp")

# simulated data
sim <-rDPM(50, priorMean_longitude = -0.04217491, priorMean_latitude = 
51.5235505, alpha=1, sigma=1, tau=3)
d <- geoData(sim$longitude, sim $latitude)
s <- geoDataSource(sim$source_lon, sim$source_lat)
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2)
m <- geoMCMC(data = d, params = p)
# raw probabilities
perspGP2(m$posteriorSurface, surface_type = "prob")
# geoprofile
perspGP2(surface = m$geoProfile, surface_type = "gp")
}
