% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.rgeoprofile_simdata}
\alias{plot.rgeoprofile_simdata}
\title{TODO}
\usage{
\method{plot}{rgeoprofile_simdata}(x, y, ...)
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}

\item{...}{TODO}
}
\description{
default plot for class rgeoprofile_simdata.
}
