% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_DIC_gelman}
\alias{plot_DIC_gelman}
\title{Plot DIC over all K}
\usage{
plot_DIC_gelman(project)
}
\arguments{
\item{project}{an RgeoProfile project, as produced by the function 
\code{rgeoprofile_project()}}
}
\description{
Plot DIC over all K for the current active parameter set.
}
