% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_expected_popsize}
\alias{plot_expected_popsize}
\title{Plot COI 95\% credible intervals}
\usage{
plot_expected_popsize(project, K = NULL)
}
\arguments{
\item{project}{an RgeoProfile project, as produced by the function 
\code{rgeoprofile_project()}}

\item{K}{which value of K to plot}
}
\description{
Plot COI 95\% credible intervals of current active set
}
\details{
TODO
}
\examples{
# TODO
}
