% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_geoprofile}
\alias{plot_geoprofile}
\title{Plot geoprofile}
\usage{
plot_geoprofile(project, K = NULL, source = NULL)
}
\arguments{
\item{project}{an RgeoProfile project, as produced by the function 
\code{rgeoprofile_project()}}

\item{K}{which value of K to plot}

\item{source}{which source to plot. If NULL then plot combined surface}
}
\description{
Plot geoprofile of selected sources.
}
\details{
TODO
}
\examples{
# TODO
}
