% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_loglike}
\alias{plot_loglike}
\title{Plot loglikelihood 95\% credible intervals}
\usage{
plot_loglike(project, K = NULL, axis_type = 1,
  connect_points = FALSE, connect_whiskers = FALSE)
}
\arguments{
\item{project}{an RgeoProfile project, as produced by the function 
\code{rgeoprofile_project()}}

\item{K}{which value of K to produce the plot for}

\item{axis_type}{how to format the x-axis. 1 = integer rungs, 2 = values of
beta, 3 = values of beta raised to the GTI power}

\item{connect_points}{whether to connect points in the middle of intervals}

\item{connect_whiskers}{whether to connect points at the ends of the whiskers}
}
\description{
Plot loglikelihood 95\% credible intervals of current active set
}
