% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lorenz}
\alias{plot_lorenz}
\title{Produce Lorenz plot of hitscores}
\usage{
plot_lorenz(hs, col = NULL)
}
\arguments{
\item{hs}{dataframe of hitscores}

\item{col}{vector of group colours. Uses \code{more_colours()} by default}
}
\description{
Produce Lorenz plot of hitscores
}
