% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_map}
\alias{plot_map}
\title{Create dynamic map}
\usage{
plot_map(map_type = 97)
}
\arguments{
\item{map_type}{an index from 1 to 137 indicating the type of base map. The
map types are taken from \code{leaflet::providers}. Defaults to "CartoDB"}
}
\description{
Create dynamic map
}
