% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_source_raw}
\alias{plot_source_raw}
\title{Plot raw source locations}
\usage{
plot_source_raw(project, K = NULL)
}
\arguments{
\item{project}{an RgeoProfile project, as produced by the function 
\code{rgeoprofile_project()}}

\item{K}{which value of K to plot}
}
\description{
Plot raw source locations.
}
\details{
TODO
}
\examples{
# TODO
}
