% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_spatial_structure}
\alias{plot_spatial_structure}
\title{Posterior allocation plot in space}
\usage{
plot_spatial_structure(project, K = NULL, pie_radius = 0.5)
}
\arguments{
\item{project}{an RgeoProfile project, as produced by the function 
\code{rgeoprofile_project()}}

\item{K}{which value of K to produce the plot for}

\item{pie_radius}{radius of pie charts}
}
\description{
Produce posterior allocation plot in space of current active set.
}
