% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_surface}
\alias{plot_surface}
\title{Plot surface of source locations}
\usage{
plot_surface(project, K = NULL, source = NULL, zlim = NULL)
}
\arguments{
\item{project}{an RgeoProfile project, as produced by the function 
\code{rgeoprofile_project()}}

\item{K}{which value of K to plot}

\item{source}{which source to plot. If NULL then plot combined surface}

\item{zlim}{z limits. If NULL then chosen automatically}
}
\description{
Plot surface of source locations
}
\details{
TODO
}
\examples{
# TODO
}
