% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print.rgeoprofile_expected_popsize_intervals}
\alias{print.rgeoprofile_expected_popsize_intervals}
\title{TODO}
\usage{
\method{print}{rgeoprofile_expected_popsize_intervals}(x, ...)
}
\arguments{
\item{x}{TODO}

\item{...}{TODO}
}
\description{
custom print function for rgeoprofile_expected_popsize_intervals.
}
