% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print.rgeoprofile_prob_surface}
\alias{print.rgeoprofile_prob_surface}
\title{TODO}
\usage{
\method{print}{rgeoprofile_prob_surface}(x, ...)
}
\arguments{
\item{x}{TODO}

\item{...}{TODO}
}
\description{
custom print function for rgeoprofile_prob_surface.
}
