% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoprofile_project.R
\name{print.rgeoprofile_project}
\alias{print.rgeoprofile_project}
\title{Custom print function for class rgeoprofile_project}
\usage{
\method{print}{rgeoprofile_project}(x, ...)
}
\arguments{
\item{x}{object of class \code{rgeoprofile_project}}

\item{...}{other arguments (ignored)}
}
\description{
Custom print function for class rgeoprofile_project, printing a summary 
  of the key elements (also equivalent to \code{summary(x)}). To do an 
  ordinary \code{print()} of all elements of the project, use the 
  \code{print_full()} function.
}
