% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print.rgeoprofile_qmatrix}
\alias{print.rgeoprofile_qmatrix}
\title{TODO}
\usage{
\method{print}{rgeoprofile_qmatrix}(x, ...)
}
\arguments{
\item{x}{TODO}

\item{...}{TODO}
}
\description{
custom print function for rgeoprofile_qmatrix.
}
