% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print.rgeoprofile_simdata}
\alias{print.rgeoprofile_simdata}
\title{TODO}
\usage{
\method{print}{rgeoprofile_simdata}(x, ...)
}
\arguments{
\item{x}{TODO}

\item{...}{TODO}
}
\description{
custom print function for rgeoprofile_simdata.
}
