% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoprofile_project.R
\name{print_full}
\alias{print_full}
\title{Ordinary print function for class rgeoprofile_project}
\usage{
print_full(x, ...)
}
\arguments{
\item{x}{object of class \code{rgeoprofile_project}}

\item{...}{other arguments passed to \code{print()}}
}
\description{
Calling \code{print()} on an object of class rgeoprofile_project results
  in custom output. This function therefore stands in for the base
  \code{print()} function, and is equivalent to running
  \code{print(unclass(x))}.
}
