% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{prob_coallocation}
\alias{prob_coallocation}
\title{Calculates the probability that two crimes are from the same source.
Also allows an optional plot showing the probabilities of allocation different
sources for each of the two selected crimes, using myMCMC$allocation.
The function returns the position of the two chosen crimes in the original
list, their lon/lat and the probability that they come from the same source.}
\usage{
prob_coallocation(crime1, crime2, new_matrix = as.matrix(myMCMC$allocation),
  offset = 0.005, plot.graph = TRUE, crime_list = crime_data)
}
\description{
Calculates the probability that two crimes are from the same source.
Also allows an optional plot showing the probabilities of allocation different
sources for each of the two selected crimes, using myMCMC$allocation.
The function returns the position of the two chosen crimes in the original
list, their lon/lat and the probability that they come from the same source.
}
\examples{
prob_coallocation()
}

