% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{raster_from_shapefile}
\alias{raster_from_shapefile}
\title{Make raster from shapefile}
\usage{
raster_from_shapefile(shp, cells_lon = 100, cells_lat = 100)
}
\arguments{
\item{shp}{shapefile to convert to raster}

\item{cells_lon}{number of cells in longitude direction}

\item{cells_lat}{number of cells in latitude direction}
}
\description{
Make raster from shapefile
}
