% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{raster_grid}
\alias{raster_grid}
\title{Make raster grid}
\usage{
raster_grid(range_lon = c(-0.2, 0), range_lat = c(51.45, 51.55),
  cells_lon = 100, cells_lat = 100)
}
\arguments{
\item{range_lon}{min and max longitude}

\item{range_lat}{min and max latitude}

\item{cells_lon}{number of cells in longitude direction}

\item{cells_lat}{number of cells in latitude direction}
}
\description{
Make raster grid
}
