% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rDPM}
\alias{rDPM}
\title{Draw from Dirichlet process mixture model}
\usage{
rDPM(n, sigma = 1, tau = 10, priorMean_longitude = -0.1277,
  priorMean_latitude = 51.5074, alpha = 1)
}
\arguments{
\item{n}{number of draws.}

\item{sigma}{standard deviation of dispersal distribution, in units of kilometres.}

\item{tau}{standard deviation of prior on source locations (i.e. average distances of sources from centre point), in units of kilometres.}

\item{priorMean_longitude}{location of prior mean on source locations in degrees longitude.}

\item{priorMean_latitude}{location of prior mean on source locations in degrees latitude.}

\item{alpha}{concentration parameter of Dirichlet process model. Large alpha implies many distinct sources, while small alpha implies only a few sources.}
}
\description{
Provides random draws from a 2D spatial Dirichlet process mixture model. Both sigma and tau are defined in units of kilometres, representing the average distance that an observations lies from a source, and the average distance that a source lies from the centre point respectively. In contrast, the location of the centre point and the locations of the final output crime sites are defined in units of degrees lat/long to facilitate spatial analysis.
}
\details{
Output includes the lat/long locations of the points drawn from the DPM model, along with the underlying group allocation (i.e. which points belong to which sources) and the lat/long locations of the sources.
}
\examples{
# produces clusters of points from sources centred on QMUL
rDPM(50, priorMean_longitude = -0.04217491, priorMean_latitude = 51.5235505, 
alpha=1, sigma=1, tau=3) 
# same, but increasing alpha to generate more clusters
rDPM(50, priorMean_longitude = -0.04217491, priorMean_latitude = 51.5235505, 
alpha=5, sigma=1, tau=3)
}
