% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RgeoProfile.R
\docType{package}
\name{RgeoProfile}
\alias{RgeoProfile}
\alias{RgeoProfile-package}
\title{RgeoProfile: Dirichlet Process Mixture (DPM) model of geographic profiling}
\description{
Carries out a DPM mixture model of geographic profiling (and associated functions) as described in Verity et al. (2014) For links to papers see http:/www.sbcs.qmul.ac.uk/staff/stevenlecomber.html or contact Steven Le Comber (s.c.lecomber@qmul.ac.uk). Created and maintained by Bob Verity (r.verity@imperial.ac.uk).
}
