% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rgeoprofile_shapefile}
\alias{rgeoprofile_shapefile}
\title{Import shapefile}
\usage{
rgeoprofile_shapefile(name)
}
\arguments{
\item{name}{name of file}
}
\description{
Import shapefile from the inst/extdata folder of this package
}
