% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{rnorm_sphere}
\alias{rnorm_sphere}
\title{Draw from bivariate normal distribution transformed to spatial coordinates}
\usage{
rnorm_sphere(n, centre_lat, centre_lon, sigma)
}
\arguments{
\item{n}{number of draws.}

\item{centre_lat}{latitude of the centre point of the normal distribution.}

\item{centre_lon}{longitude of the centre point of the normal distribution.}

\item{sigma}{standard deviation of normal distribution in km.}
}
\description{
Draw from normal distribution converted to spherical coordinate system. Points are first drawn from an ordinary cartesian 2D normal distribution. The distances to points are then assumed to be great circle distances, and are combined with a random bearing from the point {centre_lat, centre_lon} to produce a final set of lat/lon points. Note that this is not a truly spherical normal distribution, as the domain of the distribution is not the sphere - rather it is a transformation from one coordinate system to another that is satisfactory when the curvature of the sphere is not severe.
}
\examples{
rnorm_sphere(5, 51.5074, -0.1277, 1)
}
