% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate data}
\usage{
sim_data(sentinel_lon, sentinel_lat, sentinel_radius = 0.1, K = 3,
  source_lon_min = -0.2, source_lon_max = 0, source_lat_min = 51.45,
  source_lat_max = 51.55, sigma_model = "single", sigma_mean = 1,
  sigma_var = 0.1, expected_popsize = 100)
}
\arguments{
\item{sentinel_lon}{vector giving longitudes of sentinel sites}

\item{sentinel_lat}{vector giving latitudes of sentinel sites}

\item{sentinel_radius}{observation radius of the sentinel site (km)}

\item{K}{the number of sources}

\item{source_lon_min}{minimum limit on source longitudes}

\item{source_lon_max}{maximum limit on source longitudes}

\item{source_lat_min}{minimum limit on source latitudes}

\item{source_lat_max}{maximum limit on source latitudes}

\item{sigma_model}{set as "single" to use the same dispersal distance for all
sources, or "separate" to use an independently drawn dispersal distance for
each source}

\item{sigma_mean}{the prior mean of the parameter sigma (km)}

\item{sigma_var}{the prior variance of the parameter sigma (km). Set to zero 
to use a fixed distance}

\item{expected_popsize}{the expected total number of observations (observed
and unobserved) in the study area}
}
\description{
Simulate data from the same presence-absence model used in the
  inference step.
}
\examples{
# TODO
}
