% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sim_data_safe}
\alias{sim_data_safe}
\title{Simulate genetic data subject to constraints}
\usage{
sim_data_safe(..., data_format = "biallelic", no_invariant_loci = TRUE,
  no_missing_samples = TRUE, no_missing_loci = TRUE, max_attempts = 1000)
}
\arguments{
\item{...}{TODO}

\item{data_format}{TODO}

\item{no_invariant_loci}{TODO}

\item{no_missing_samples}{TODO}

\item{no_missing_loci}{TODO}

\item{max_attempts}{TODO}
}
\description{
TODO - text
}
\details{
TODO
}
\examples{
# TODO
}
