% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoprofile_project.R
\name{summary.rgeoprofile_project}
\alias{summary.rgeoprofile_project}
\title{Print summary for class rgeoprofile_project}
\usage{
summary.rgeoprofile_project(object, ...)
}
\arguments{
\item{object}{object of class \code{rgeoprofile_project}}

\item{...}{other arguments (ignored)}
}
\description{
Overload summary function for class rgeoprofile_project
}
