% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WaterPumps}
\alias{WaterPumps}
\title{Cholera sources}
\format{an object of class \code{list}.}
\usage{
data(WaterPumps)
}
\description{
Locations of the neighbourhood water pumps near the site of the 1854 cholera outbreak. The Broad Street pump was the source of the outbreak.
}
\examples{
plot(Cholera$latitude, Cholera$longitude, xlab = "lon", ylab = "lat")
points(WaterPumps$latitude, WaterPumps$longitude,  pch = 15, col = "blue")

}
\keyword{datasets}
