% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoData}
\alias{geoData}
\title{Create Rgeoprofile data object}
\usage{
geoData(longitude = NULL, latitude = NULL)
}
\arguments{
\item{longitude}{the locations of the observed data in degrees longitude.}

\item{latitude}{the locations of the observed data in degrees latitude.}
}
\description{
Simple function that ensures that input data is in the correct format required by Rgeoprofile. Takes longitude and latitude as input vectors and returns these same values in list format. If no values are input then default values are used.
}
\examples{
# John Snow cholera data
data(Cholera)
geoData(Cholera$longitude, Cholera$latitude)

# simulated data
sim <-rDPM(50, priorMean_longitude = -0.04217491, priorMean_latitude = 
51.5235505, alpha=1, sigma=1, tau=3)
geoData(sim$longitude, sim $latitude)
}
