% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoData}
\alias{geoData}
\title{Create Rgeoprofile data object}
\usage{
geoData(longitude = NULL, latitude = NULL)
}
\arguments{
\item{longitude}{The longitudinal positions of the observed data}

\item{latitude}{The latitudinal positions of the observed data}
}
\description{
This function can be used to generate a dummy data object in the format used by the Rgeoprofile MCMC.
}
\examples{
geoData()
}

