% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoDataCheck}
\alias{geoDataCheck}
\title{Check data}
\usage{
geoDataCheck(data, silent = FALSE)
}
\arguments{
\item{data}{A list of data}
}
\description{
Check that all data for use in Rgeoprofile MCMC is in the correct format.
}
\examples{
# tester
geoDataCheck(myData)
}
\keyword{bob}

