% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoDataCheck}
\alias{geoDataCheck}
\title{Check data}
\usage{
geoDataCheck(data, silent = FALSE)
}
\arguments{
\item{data}{a data list object, as defined by geoData().}

\item{silent}{whether to report if data passes checks to console.}
}
\description{
Check that all data for use in Rgeoprofile MCMC is in the correct format.
}
\examples{
# John Snow cholera data
d <- geoData(Cholera$longitude, Cholera$latitude)
geoDataCheck(d)

# simulated data
sim <-rDPM(50, priorMean_longitude = -0.04217491, priorMean_latitude = 
51.5235505, alpha=1, sigma=1, tau=3)
d <- geoData(sim$longitude, sim $latitude)
geoDataCheck(d)
}
