% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoDataCheck}
\alias{geoDataCheck}
\title{Check data}
\usage{
geoDataCheck(data, silent = FALSE)
}
\arguments{
\item{data}{A data list object, as defined by geoData()}

\item{silent}{Whether to report passing check to console}
}
\description{
Check that all data for use in Rgeoprofile MCMC is in the correct format.
}
\examples{
# tester
geoDataCheck(myData)
}

