% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoDataSource}
\alias{geoDataSource}
\title{Create sources data object in same format as observations}
\usage{
geoDataSource(source_longitude = NULL, source_latitude = NULL)
}
\arguments{
\item{longitude}{the longitudinal positions of the potential sources.}

\item{latitude}{the latitudinal positions of the potential sources.}
}
\description{
Simple function that ensures that sources are in the correct format required by Rgeoprofile. Takes longitude and latitude as input vectors and returns these same values in list format. If no values are input then default values are used.
}
\examples{
# john snow cholera data
data(WaterPumps)
geoDataSource(WaterPumps[,1], WaterPumps[,2])

# simulated data
sim <-rDPM(50, priorMean_longitude = -0.04217491, priorMean_latitude = 
51.5235505, alpha=1, sigma=1, tau=3)
geoDataSource(sim$source_longitude, sim$source_latitude)
}
