% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoDataSource}
\alias{geoDataSource}
\title{Create sources data object in same format}
\usage{
geoDataSource(source_longitude = NULL, source_latitude = NULL)
}
\description{
Create sources data object in same format
}

