% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoMCMC}
\alias{geoMCMC}
\title{MCMC under Rgeoprofile model}
\usage{
geoMCMC(data, params)
}
\arguments{
\item{data}{input data in the format defined by geoData}

\item{params}{input parameters in the format defined by geoParams}
}
\description{
This function carries out the main MCMC under the Rgeoprofile model.
}
\examples{
geoMCMC()
}

