% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoMCMC}
\alias{geoMCMC}
\title{MCMC under Rgeoprofile model}
\usage{
geoMCMC(data, params)
}
\arguments{
\item{data}{input data in the format defined by geoData().}

\item{params}{input parameters in the format defined by geoParams().}
}
\description{
This function carries out the main MCMC under the Rgeoprofile model.
}
\examples{
data(Cholera)
d <- geoData(Cholera[,1],Cholera[,2])

p <- geoParams(data=d,sigma_mean=1.0,sigma_squared_shape=2,samples=20000,chains=10,burnin=1000,priorMean_longitude=mean(dat$longitude),
priorMean_latitude=mean(dat$latitude),guardRail=0.1)
m <- geoMCMC(datda=d,params=p)
}
