% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoParams}
\alias{geoParams}
\title{Create Rgeoprofile parameters object}
\usage{
geoParams(data = NULL, sigma = 1, priorMean_longitude = 0,
  priorMean_latitude = 0, priorSD = 10, alpha_shape = 0.1,
  alpha_rate = 0.1, chains = 10, burnin = 500, samples = 5000,
  burnin_printConsole = 100, samples_printConsole = 1000,
  longitude_minMax = c(-10, 10), latitude_minMax = c(-10, 10),
  longitude_cells = 500, latitude_cells = 500)
}
\arguments{
\item{sigma}{The standard deviation of the dispersal distribution}

\item{priorMean_longitude}{The position (longitude) of the mean of the prior distribution}

\item{priorMean_latitude}{The position (latitude) of the mean of the prior distribution}
}
\description{
This function can be used to generate parameters in the format used by the Rgeoprofile MCMC. Particular parameter values can be specified as input arguments.
}
\examples{
geoParams()
}
\keyword{bob}

