% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoParams}
\alias{geoParams}
\title{Create Rgeoprofile parameters object}
\usage{
geoParams(data = NULL, sigma_mean = 1, sigma_var = NULL,
  sigma_squared_shape = NULL, sigma_squared_rate = NULL,
  priorMean_longitude = NULL, priorMean_latitude = NULL, tau = NULL,
  alpha_shape = 0.1, alpha_rate = 0.1, chains = 10, burnin = 500,
  samples = 5000, burnin_printConsole = 100, samples_printConsole = 1000,
  longitude_minMax = NULL, latitude_minMax = NULL, longitude_cells = 500,
  latitude_cells = 500, guardRail = 0.05)
}
\arguments{
\item{data}{observations in the format defined by geoData().}

\item{sigma_mean}{the mean of the prior on sigma (sigma = standard deviation of the dispersal distribution).}

\item{sigma_var}{the variance of the prior on sigma.}

\item{sigma_squared_shape}{as an alternative to defining the prior mean and variance of sigma, it is possible to directly define the parameters of the inverse-gamma prior on sigma^2. This is the shape parameter of the inverse-gamma prior.}

\item{sigma_squared_rate}{this is the rate parameter of the inverse-gamma prior on sigma^2.}

\item{priorMean_longitude}{the position of the prior mean on source locations in degrees longitude.}

\item{priorMean_latitude}{the position of the prior mean on source locations in degrees latitude.}

\item{tau}{the standard deviation of the normal prior on source locations, i.e. how far we expect sources to lie from the centre. Leave as NULL to use default value, in which case tau is set equal to the maximum distance of any observation from the prior mean.}

\item{alpha_shape}{shape parameter of the gamma prior on the parameter alpha.}

\item{alpha_rate}{rate parameter of the gamma prior on the parameter alpha.}

\item{chains}{number of MCMC chains to use in the burn-in step.}

\item{burnin}{number of burn-in iterations to be discarded at start of MCMC.}

\item{samples}{number of sampling iterations. These iterations are used to generate final posterior distribution.}

\item{burnin_printConsole}{how frequently (in iterations) to report progress to the console during the burn-in phase.}

\item{samples_printConsole}{how frequently (in iterations) to report progress to the console during the sampling phase.}

\item{longitude_minMax}{vector containing minimum and maximum longitude over which to generate geoprofile.}

\item{latitude_minMax}{vector containing minimum and maximum latitude over which to generate geoprofile.}

\item{longitude_cells}{number of cells in the final geoprofile (longitude direction). Higher values generate smoother distributions, but take longer to run.}

\item{latitude_cells}{number of cells in the final geoprofile (latitude direction). Higher values generate smoother distributions, but take longer to run.}
}
\description{
This function can be used to generate parameters in the format required by other Rgeoprofile functions. Any parameter value can be specified as an input argument to this function. Alternatively, if data is input as an argument then some parameters can take default values directly from the data.
}
\examples{
data(Cholera)
dat <- geoData(Cholera[,1],Cholera[,2])
# using weak prior on sigma
geoParams(dat,sigma_mean=1.0,sigma_squared_shape=2,samples=20000,chains=10,burnin=1000,priorMean_longitude=mean(dat$longitude),priorMean_latitude=mean(dat$latitude),guardRail=0.05)
# using fixed sigma
geoParams(dat,sigma_mean=1.0,sigma_var=0,samples=20000,chains=10,burnin=1000,priorMean_longitude=mean(dat$longitude),priorMean_latitude=mean(dat$latitude),guardRail=0.05)
}
