% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoParamsCheck}
\alias{geoParamsCheck}
\title{Check parameters}
\usage{
geoParamsCheck(params, silent = FALSE)
}
\arguments{
\item{params}{A list of parameters}

\item{silent}{Whether to report passing check to console}
}
\description{
Check that all parameters for use in Rgeoprofile MCMC are OK.
}
\examples{
# tester
geoParamsCheck(myParams)
}

