% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoParamsCheck}
\alias{geoParamsCheck}
\title{Check parameters}
\usage{
geoParamsCheck(params, silent = FALSE)
}
\arguments{
\item{params}{a list of parameters, as defined by geoParams().}

\item{silent}{whether to report passing check to console.}
}
\description{
Check that all parameters for use in Rgeoprofile MCMC are in the correct format.
}
\examples{
# John Snow cholera data
data(Cholera)
d <- geoData(Cholera[,1],Cholera[,2])
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2)
geoParamsCheck(p)

# simulated data
sim <-rDPM(50, priorMean_longitude = -0.04217491, priorMean_latitude = 
51.5235505, alpha=1, sigma=1, tau=3)
d <- geoData(Cholera[,1], Cholera[,2])
p <- geoParams(data = d, sigma_mean = 1.0, sigma_var=0)
geoParamsCheck(p)
}
