% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoPlotAllocation}
\alias{geoPlotAllocation}
\title{Plot posterior allocation}
\usage{
geoPlotAllocation(allocation, colours = "default", barBorderCol = "white",
  barBorderWidth = 0.25, mainBorderCol = "black", mainBorderWidth = 2,
  yTicks_on = TRUE, yTicks = seq(0, 1, 0.2), xlab = "",
  ylab = "posterior allocation", mainTitle = "", names = NA,
  names_size = 1, xTicks_on = FALSE, xTicks_size = 1, orderBy = "group")
}
\arguments{
\item{MCMCoutput}{output generated from an MCMC.}
}
\description{
Produces plot of posterior allocation from output of MCMC.
}
\examples{
# john snow cholera data
data(Cholera)
d <- geoData(Cholera[,1],Cholera[,2])
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2, samples = 20000, 
chains = 10, burnin = 1000, priorMean_longitude = mean(d$longitude), 
priorMean_latitude = mean(d$latitude), guardRail = 0.1)
m <- geoMCMC(data = d, params = p)
geoPlotAllocation(m$allocation)

# simulated data
sim <-rDPM(50, priorMean_longitude = -0.04217491, priorMean_latitude = 
51.5235505, alpha=1, sigma=1, tau=3)
d <- geoData(sim$longitude, sim $latitude)
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2, samples = 20000, 
chains = 10, burnin = 1000, priorMean_longitude = mean(d$longitude), 
priorMean_latitude = mean(d$latitude), guardRail = 0.1)
m <- geoMCMC(data = d, params = p)
geoPlotAllocation(m$allocation)
}
