% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoPlotAllocation}
\alias{geoPlotAllocation}
\title{Plot posterior allocation}
\usage{
geoPlotAllocation(allocation, colours = "default", barBorderCol = "white",
  barBorderWidth = 0.25, mainBorderCol = "black", mainBorderWidth = 2,
  yTicks_on = TRUE, yTicks = seq(0, 1, 0.2), xlab = "",
  ylab = "posterior allocation", mainTitle = "", names = NA,
  names_size = 1, xTicks_on = FALSE, xTicks_size = 1, orderBy = "group")
}
\arguments{
\item{MCMCoutput}{output generated from an MCMC.}
}
\description{
Produces plot of posterior allocation from output of MCMC.
}
\examples{
geoPlotAllocation(MCMCoutput)
}
