% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoPlotLorenz}
\alias{geoPlotLorenz}
\title{Produce Lorenz Plot}
\usage{
geoPlotLorenz(hit_scores, crimeNumbers = NULL, suspects_col = "red",
  crimes_col = "blue")
}
\arguments{
\item{hit_scores}{object in the format defined by geoReportHitscores().}

\item{crimeNumbers}{optional vector with numbers of crimes per suspect site.}
}
\description{
Produces a Lorenz plot showing the proportion of suspect sites or cimes identified as a function of area and calculates
the corresponding Gini coefficient using trapezoid rule.
Also allows an optional vector called crimeNumbers with numbers of crimes per suspect site; tthe length of this vector
should equal the number of suspect sites. If this is present, the function calculates and returns the Gini coefficient 
based on the number of crimes; otherwise, this is calculated based on the number of suspect sites.
}
\examples{
data(Cholera)
d <- geoData(Cholera[,1],Cholera[,2])
data(WaterPumps)
s <- geoDataSource(WaterPumps[,1], WaterPumps[,2])
p <- geoParams(data=d,sigma_mean=1.0,sigma_squared_shape=2,samples=20000,chains=10,burnin=1000,priorMean_longitude=mean(dat$longitude),priorMean_latitude=mean(dat$latitude),guardRail=0.1)
m <- geoMCMC(data=d,params=p)
gp <- geoProfile(surface=m$surface)
hs <- geoReportHitscores(params=p,source_data=s,surface=m$surface)
geoPlotLorenz(hit_scores=hs)
geoPlotLorenz(hit_scores=hs,crimeNumbers=c(10,1,2,3,1,1,1))
}
