% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{geoPlotMap}
\alias{geoPlotMap}
\title{Plot a map and overlay data and/or geoprofile}
\usage{
geoPlotMap(params, data = NULL, source = NULL, surface = NULL,
  surfaceCols = NULL, zoom = NULL, latLimits = NULL, lonLimits = NULL,
  mapSource = "google", mapType = "hybrid", opacity = 0.6,
  plotContours = TRUE, breakPercent = seq(0, 100, l = 11),
  contourCol = "grey50", smoothScale = TRUE, crimeCex = 1.5,
  crimeCol = "red", crimeBorderCol = "white", crimeBorderWidth = 0.5,
  sourceCex = 1.5, sourceCol = "blue", gpLegend = TRUE)
}
\arguments{
\item{params}{parameters list in the format defined by geoParams().}

\item{data}{data object in the format defined by geoData().}

\item{source}{potential sources object in the format defined by geoDataSource().}

\item{surface}{a surface to overlay onto the map, typically a geoprofile obtained from the output of geoMCMC().}

\item{surfaceCols}{TODO.}

\item{zoom}{zoom level of map. If NULL then choose optimal zoom from params.}

\item{latLimits}{TODO}

\item{lonLimits}{TODO}

\item{mapSource}{which online source to use when downloading the map. Options include Google Maps ("google"), OpenStreetMap ("osm"), Stamen Maps ("stamen") and CloudMade maps ("cloudmade").}

\item{mapType}{the specific type of map to plot. Options available are "terrain", "satellite", "roadmap" and "hybrid" (google maps), "terrain-background", "terrain", "watercolor" and "toner" (stamen maps) or a positive integer for cloudmade maps (see ?get_cloudmademap from the package ggmap for details).}

\item{opacity}{value between 0 and 1 givin the opacity of surface colours.}

\item{plotContours}{whether or not to add contours to the surface plot.}

\item{breakPercent}{vector of values between 0 and 100 describing where in the surface contours appear.}

\item{contourCol}{TODO}

\item{smoothScale}{TODO}

\item{crimeCex}{relative size of symbols showing crimes.}

\item{crimeCol}{colour of crime symbols.}

\item{crimeBorderCol}{border colour of crime symbols.}

\item{crimeBorderWidth}{width of border of crime symbols.}

\item{sourceCex}{relative size of symbols showing suspect sites.}

\item{sourceCol}{colour of suspect sites symbols.}

\item{gpLegend}{whether or not to add legend to plot.}
}
\description{
Plots geoprofile on map, with various customisable options.
}
\examples{
# London example data
d <- LondonExample_crimes
s <- LondonExample_sources
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2)
m <- geoMCMC(data = d, params = p)
# produce simple map
geoPlotMap(params = p, data = d, source = s, surface = m$geoProfile,
                breakPercent = seq(0, 50, 5), mapType = "hybrid",
                crimeCol = "black", crimeCex = 2, sourceCol = "red", sourceCex = 2)

# John Snow cholera data
d <- Cholera
s <- WaterPumps
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2)
m <- geoMCMC(data = d, params = p, lambda=0.05)
# produce simple map
geoPlotMap(params = p, data = d, source = s, surface = m$geoProfile,
                breakPercent = seq(0, 50, 5), mapType = "hybrid",
                crimeCol = "black", crimeCex = 2, sourceCol = "red", sourceCex = 2)

# simulated data
sim <-rDPM(50, priorMean_longitude = -0.04217491, priorMean_latitude = 
51.5235505, alpha=1, sigma=1, tau=3)
d <- geoData(sim$longitude, sim $latitude)
s <- geoDataSource(sim$source_lon, sim$source_lat)
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2)
m <- geoMCMC(data = d, params = p)
# change colour palette, map type, opacity and range of geoprofile and omit legend
geoPlotMap(params = p, data = d, source = s, surface = m$geoProfile,
                breakPercent = seq(0, 30, 5), mapType = "terrain", 
                surfaceCols = c("blue","white"), crimeCol = "black", 
                crimeBorderCol = "white",crimeCex = 2, sourceCol = "red", sourceCex = 2,
                opacity = 0.7, gpLegend = FALSE)
}
