% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoPlotSigma}
\alias{geoPlotSigma}
\title{Plot prior and posterior distributions of sigma.}
\usage{
geoPlotSigma(params, sigma = NULL, plotMax = NULL)
}
\arguments{
\item{params}{a list of parameters as defined by geoParams().}

\item{sigma}{a vector of posterior draws of sigma. Leave as NULL to plot prior only.}

\item{plotMax}{maximum x-axis range to plot. Leave as NULL to use default settings.}
}
\description{
Plot prior distribution of sigma as defined by current parameter values. Can optionally overlay a kernel density plot of posterior draws of sigma.
}
\examples{
data(Cholera)
dat <- geoData(Cholera[,1],Cholera[,2])
p <- geoParams(dat,sigma_mean=1.0,sigma_squared_shape=2,samples=20000,chains=10,burnin=1000,priorMean_longitude=mean(dat$longitude)
,priorMean_latitude=mean(dat$latitude),guardRail=0.05)
m <- geoMCMC(dat,p)
geoPlotSigma(p,m$sigma)
}
