% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoPlotSigma}
\alias{geoPlotSigma}
\title{Plot prior and posterior distributions of sigma.}
\usage{
geoPlotSigma(params, sigma = NULL, plotMax = NULL)
}
\arguments{
\item{params}{a list of parameters as defined by geoParams().}

\item{sigma}{a vector of posterior draws of sigma. Leave as NULL to plot prior only.}

\item{plotMax}{maximum x-axis range to plot. Leave as NULL to use default settings.}
}
\description{
Plot prior distribution of sigma as defined by current parameter values. Can optionally overlay a kernel density plot of posterior draws of sigma.
}
\examples{
myData <- geoData()
myParams <- geoParams(myData, sigma_var=1)
geoPlotSigma(myParams)
}

