% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoProfile}
\alias{geoProfile}
\title{Calculate geoprofile from surface}
\usage{
geoProfile(surface)
}
\arguments{
\item{surface}{matrix to convert to geoprofile}
}
\description{
Converts surface to rank order geoprofile.
}
\examples{
data(Cholera)
dat <- geoData(Cholera[,1],Cholera[,2])
p <- geoParams(dat,sigma_mean=1.0,sigma_squared_shape=2,samples=20000,chains=10,burnin=1000,priorMean_longitude=mean(dat$longitude),priorMean_latitude=mean(dat$latitude),guardRail=0.05)
m <- geoMCMC(dat,p)
geoProfile(m$surface)
}
