% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoProfile}
\alias{geoProfile}
\title{Calculate geoprofile from surface}
\usage{
geoProfile(surface)
}
\arguments{
\item{surface}{matrix to convert to geoprofile}
}
\description{
Converts surface to rank order geoprofile.
}
\examples{
data(Cholera)
d <- geoData(Cholera[,1],Cholera[,2])
p <- geoParams(data=d,sigma_mean=1.0,sigma_squared_shape=2,samples=20000,chains=10,
burnin=1000,priorMean_longitude=mean(d$longitude),
priorMean_latitude=mean(d$latitude),guardRail=0.1)
m <- geoMCMC(data=d,params=p)
geoProfile(m$surface)
}
