% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoProfile}
\alias{geoProfile}
\title{Calculate geoprofile from surface}
\usage{
geoProfile(surface)
}
\arguments{
\item{surface}{matrix to convert to geoprofile}
}
\description{
Converts surface to rank order geoprofile.
}
\examples{
# john snow cholera data
data(Cholera)
d <- geoData(Cholera[,1],Cholera[,2])
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2, samples = 20000, 
chains = 10, burnin = 1000, priorMean_longitude = mean(d$longitude), 
priorMean_latitude = mean(d$latitude), guardRail = 0.1)
m <- geoMCMC(data = d, params = p)
gp <- geoProfile(m$surface)

# simulated data
sim <-rDPM(50, priorMean_longitude = -0.04217491, priorMean_latitude = 
51.5235505, alpha=1, sigma=1, tau=3)
d <- geoData(sim$longitude, sim $latitude)
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2, samples = 20000, 
chains = 10, burnin = 1000, priorMean_longitude = mean(d$longitude), 
priorMean_latitude = mean(d$latitude), guardRail = 0.1)
m <- geoMCMC(data = d, params = p)
gp <- geoProfile(m$surface)
}
