% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoProfile}
\alias{geoProfile}
\title{Calculate geoprofile from surface}
\usage{
geoProfile(z)
}
\arguments{
\item{z}{matrix to convert to geoprofile}
}
\description{
Converts surface to rank order geoprofile.
}
\examples{
geoProfile(MCMCoutput$geoSurface)
}
\keyword{bob}

