% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoQuickPlot}
\alias{geoQuickPlot}
\title{Create quick geoprofile plot}
\usage{
geoQuickPlot(params, surface = NULL, data = NULL, zoom = "auto",
  source = "google", maptype = "hybrid", breakPercent = seq(0, 10, l =
  11), plotContours = TRUE, data_fillCol = "black",
  data_borderCol = "white")
}
\arguments{
\item{surface}{}
}
\description{
Creates quick geoprofile plot, choosing some parameters automatically.
}
\examples{
geoQuickPlot(surface)
}
\keyword{bob}

