% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoQuickPlot}
\alias{geoQuickPlot}
\title{Create quick geoprofile plot}
\usage{
geoQuickPlot(params, surface = NULL, data = NULL, zoom = "auto",
  source = "google", maptype = "hybrid", breakPercent = seq(0, 100, l =
  11), contour_cols = c("red", "orange", "yellow", "white"),
  plotContours = TRUE, crimeCol = "red", crimePch = 16, crimeCex = 1,
  CrimeBorderCol = "white", sourceCol = "blue", sourcePch = 15,
  sourceCex = 1, SourceBorderCol = "white", source_data = source_data)
}
\arguments{
\item{surface}{some text}
}
\description{
Creates quick geoprofile plot, choosing some parameters automatically. Now plots sources too (SLC).
}
\examples{
geoQuickPlot(surface)
}

