% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoReportHitscores}
\alias{geoReportHitscores}
\title{Calculate hitscores}
\usage{
geoReportHitscores(params, source_data, surface)
}
\arguments{
\item{params}{some text}

\item{source_data}{some text}

\item{surface}{some text}
}
\description{
Calculate hitscores
}
\details{
Calculate hitscores of the potential sources based on the final geoprofile surface.
}
\examples{
geoReportHitscores(params,source_data,surface)
}

