% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoReportHitscores}
\alias{geoReportHitscores}
\title{Calculate hitscores}
\usage{
geoReportHitscores(params, source, surface)
}
\arguments{
\item{params}{TODO}

\item{source}{longitude and latitude of one or more source locations in the format defined by geoDataSource().}

\item{surface}{TODO}
}
\description{
Calculate hitscores of the potential sources for a given surface (usually the geoprofile).
}
\examples{
# TODO

}
