% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoReportHitscores}
\alias{geoReportHitscores}
\title{Calculate hitscores}
\usage{
geoReportHitscores(params, source_data, surface)
}
\arguments{
\item{params}{some text}

\item{source_data}{some text}

\item{surface}{some text}
}
\description{
Calculate hitscores of the potential sources based on the final geoprofile surface.
}
\examples{
data(Cholera)
d <- geoData(Cholera[,1],Cholera[,2])
data(WaterPumps)
s <- geoDataSource(WaterPumps[,1], WaterPumps[,2])
p <- geoParams(data=d,sigma_mean=1.0,sigma_squared_shape=2,samples=20000,chains=10,
burnin=1000,priorMean_longitude=mean(dat$longitude),
priorMean_latitude=mean(dat$latitude),guardRail=0.1)
m <- geoMCMC(data=d,params=p)
gp <- geoProfile(surface=m$surface)
hs <- geoReportHitscores(params=p,source_data=s,surface=m$surface)
}
