% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoReportHitscores}
\alias{geoReportHitscores}
\title{Calculate hitscores}
\usage{
geoReportHitscores(params, source_data, surface)
}
\arguments{
\item{mcmc}{stored output obtained by running geoMCMC().}

\item{source}{longitude and latitude of one or more source locations in the format defined by geoDataSource().}
}
\description{
Calculate hitscores of the potential sources for a given surface (usually the geoprofile).
}
\examples{
# John Snow cholera data
data(Cholera)
d <- geoData(Cholera[,1],Cholera[,2])
data(WaterPumps)
s <- geoDataSource(WaterPumps[,1], WaterPumps[,2])
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2, samples = 20000, 
chains = 10, burnin = 1000, priorMean_longitude = mean(d$longitude), 
priorMean_latitude = mean(d$latitude), guardRail = 0.1)
m <- geoMCMC(data = d, params = p)
gp <- m$geoProfile
geoPlotMap(params = p, data = d, source = s, breakPercent = seq(0, 50, 5), mapType = "hybrid",
contourCols = c("red", "orange", "yellow", "white"), crimeCol = "black", crimeBorderCol = "white", 
crimeCex = 2, sourceCol = "red", sourceCex = 2, surface = gp)
hs <- geoReportHitscores(params=p,source_data=s,surface=m$surface)

# simulated data
sim <-rDPM(50, priorMean_longitude = -0.04217491, priorMean_latitude = 
51.5235505, alpha=1, sigma=1, tau=3)
d <- geoData(sim$longitude, sim $latitude)
s <- geoDataSource(sim$source_lon, sim$source_lat)
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2, samples = 20000, 
chains = 10, burnin = 1000, priorMean_longitude = mean(d$longitude), 
priorMean_latitude = mean(d$latitude), guardRail = 0.1)
gp <- m$geoProfile
geoPlotMap(params = p, data = d, source = s,breakPercent = seq(0, 30, 5), mapType = "terrain", 
contourCols=c("blue","white"),crimeCol="black", crimeBorderCol="white",crimeCex=2,
sourceCol = "red", sourceCex = 2, surface = gp, transparency = 0.7)
hs <- geoReportHitscores(params=p,source_data=s,surface=m$surface)

}
