% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{geoSmooth}
\alias{geoSmooth}
\title{Produce a smooth surface using 2D kernel density smoothing}
\usage{
geoSmooth(longitude, latitude, breaks_lon, breaks_lat, lambda = NULL)
}
\arguments{
\item{longitude}{longitude of input points}

\item{latitude}{latitude of input points}

\item{breaks_lon}{positions of longitude breaks}

\item{breaks_lat}{positions of latitude breaks}

\item{lambda}{bandwidth to use in posterior smoothing. If NULL then optimal bandwidth is chosen automatically by maximum-likelihood.}
}
\description{
Takes lon/lat coordinates, bins in two dimensions, and smooths using kernel density smoothing. Kernel densities are computed using the fast Fourier transform method, which is many times faster than simple summation when using a large number of points. Each Kernel is student's-t distributed with 3 degrees of freedom, and scaled by the bandwidth lambda. If lambda is set to \code{NULL} then the optimal value of lambda is chosen automatically using the leave-one-out maximum likelihood method.
}
\examples{
# TODO
}
