% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoSmooth}
\alias{geoSmooth}
\title{kernel density smoothing of posterior distribution}
\usage{
geoSmooth(data, params, MCMCoutput, lambda = 1, df = 3)
}
\arguments{
\item{data}{input data in the format defined by geoData().}

\item{params}{input parameters in the format defined by geoParams().}

\item{MCMCoutput}{stored output of the MCMC obtained by running geoMCMC().}

\item{lambda}{scale of smoothing kernel, relative to an ordinary Student's t distribution in which lambda=1. When appled to geoMCMC() output this parameter is in units of km.}

\item{df}{degrees of freedom of Student's t smoothing kernel.}
}
\description{
Can be used to perform kernel density smoothing of the "surface_raw" object output from geoMCMC(). Note that geoMCMC() performs this smoothing already using a maximum-likelihood estimate of lambda and outputs it to the "surface" object, and so this function is only needed when a custom level of smoothing is required. The kernel used is a Student's t distribution with a user-defined scale (1 = ordinary Student's t distribution) and degrees of freedom.
}
\examples{
myData <- geoData()
myParams <- geoParams(myData, sigma_var=1)
myMCMC <- geoMCMC(myData, myParams)
mySurface <- geoSmooth(myData, myParams, myMCMC, lambda=0.5, df=2)
}
