% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoSurface}
\alias{geoSurface}
\title{Create geoprofile plot object}
\usage{
geoSurface(longitude_midpoints = 1:ncol(surface),
  latitude_midpoints = 1:nrow(surface), surface)
}
\arguments{
\item{surface}{}
}
\description{
Creates geoprofile plotting object, for use with other ggplot2 elements.
}
\examples{
geoSurface(surface)
}
\keyword{bob}

