% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{GPshapefile}
\alias{GPshapefile}
\title{Incorporate shapefile information into a geoprofile}
\usage{
GPshapefile(probSurface, params, shapefile, masterProj, scaleValue,
  excludeShapefile = FALSE)
}
\arguments{
\item{probSurface}{the original geoprofile, usually the object $posteriorSurface produced by geoMCMC().}

\item{params}{an object produced by geoParams().}

\item{shapefile}{the SpatialPolygonsDataFrame object to include.}

\item{masterProj}{the projection to use, eg "+proj=longlat +datum=WGS84".}

\item{scaleValue}{value by which probabilities should be multiplied inside (or outside, depending on excludeShapefile) the shapefile. Set to zero to exclude completely.}

\item{excludeShapefile}{must be set to TRUE (multiply areas outside the shapefile by scaleValue) or FALSE (multiply areas inside the shapefile by scaleValue).}
}
\description{
This function allows information from a shapefile to be incorporated within the geoprofile. For example, we might wish to exclude areas not on land, or weight the probabilities within a specific postcode differently. The shapefile used should be a SpatialPolygonsDataFrame as produced by the package sp.
}
\examples{
# to come
}
