% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LondonExample_crimes}
\alias{LondonExample_crimes}
\title{Example London crimes}
\format{an object of class \code{list}.}
\usage{
data(LondonExample_crimes)
}
\description{
Dummy crime sites simulated using rDPM() and centred on Queen Mary University of London.
}
\examples{
# plot(LondonExample_crimes$latitude, LondonExample_crimes$longitude, xlab = "lon", ylab = "lat")

}
\keyword{datasets}
