% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LondonExample_sources}
\alias{LondonExample_sources}
\title{TODO}
\format{an object of class \code{list}.}
\usage{
data(LondonExample_sources)
}
\description{
TODO
}
\examples{
# TODO

}
\keyword{datasets}
