% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{new_set}
\alias{new_set}
\title{Create new parameter set}
\usage{
new_set(project, spatial_prior = NULL, sentinel_radius = 0.2,
  sigma_model = "single", sigma_prior_mean = 1, sigma_prior_sd = 1,
  expected_popsize_prior_mean = 100, expected_popsize_prior_sd = 10,
  name = "(no name)")
}
\arguments{
\item{project}{an rgeoprofile_project, as produced by the function 
\code{rgeoprofile_project()}}

\item{spatial_prior}{a raster file defining the spatial prior. Precision
values are taken from this raster if it is defined}

\item{sentinel_radius}{the observation radius of sentinel sites}

\item{sigma_model}{set as \code{"single"} to assume the same dispersal
distance for all sources, or \code{"independent"} to assume an
independently drawn dispersal distance for each source}

\item{sigma_prior_mean}{the prior mean of the parameter sigma (km)}

\item{sigma_prior_sd}{the prior standard deviation of the parameter sigma 
(km). Set to 0 to use a fixed value for sigma (fixed at
\code{sigma_prior_mean})}

\item{expected_popsize_prior_mean}{the prior mean of the expected total
population size}

\item{expected_popsize_prior_sd}{the prior standard deviation of the expected
total population size. Set to 0 to use a fixed value (fixed at
\code{expected_popsize_prior_mean}), or set to -1 to use an improper,
infinitely diffuse prior}

\item{name}{an optional name for the parameter set}
}
\description{
Create a new parameter set within an \code{rgeoprofile_project}. The new 
  parameter set becomes the active set once created.
}
