% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{rDPM}
\alias{rDPM}
\title{Draw from Dirichlet process mixture model}
\usage{
rDPM(n, sigma = 0.01, priorMean_longitude = -0.1277,
  priorMean_latitude = 51.5074, priorSD = 0.03, alpha = 1)
}
\arguments{
\item{n}{number of draws}

\item{sigma}{standard deviation of dispersal distribution, in units of degrees}
}
\description{
Provides random draws from a 2D Dirichlet process mixture model. Coordinates are defined in units of degrees latitude and longitude to facilitate spatial analysis.
}
\examples{
rDPM(10)
}

