% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{rDPM}
\alias{rDPM}
\title{Draw from Dirichlet process mixture model}
\usage{
rDPM(n, sigma = 1, priorMean_x = 0, priorMean_y = 0, priorVar = 10,
  alpha = 1)
}
\arguments{
\item{n}{The number of draws}
}
\description{
This function provides draws from a 2D Dirichlet process mixture model.
}
\examples{
rDPM(10)
}
\keyword{bob}

