% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{ringHS}
\alias{ringHS}
\title{Calculates hit scores for a ring-search strategy (ie searching in an expanding radius out from the crimes).
Also plots the crimes and sources with merged polygons showing these (merged and clipped) rings}
\usage{
ringHS(my_crime_data = crime_data, my_source_data = source_data,
  buffer_radiuses = c(1000, 2000, 5000))
}
\description{
Calculates hit scores for a ring-search strategy (ie searching in an expanding radius out from the crimes).
Also plots the crimes and sources with merged polygons showing these (merged and clipped) rings
}
\examples{
ringHS()
}

