% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{ringHS}
\alias{ringHS}
\title{Calculate and plot ring search strategy}
\usage{
ringHS(data, source, buffer_radii = c(1000, 2000, 5000))
}
\arguments{
\item{data}{some text}

\item{source}{some text}

\item{buffer_radii}{some text}
}
\description{
Calculates hit scores for a ring-search strategy (ie searching in an expanding radius out from the crimes).
Also plots the crimes and sources with merged polygons showing these (merged and clipped) rings
}
\examples{
ringHS()
}
